pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi-mail>=1.6.2",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "ollama>=0.6.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-socketio>=5.16.1",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.41.0",
]
